Remove single quotes
We do not put items in single quotes. I put the term in bold where it is first mentioned, as bold is the way we refer to what customers can see as they use a component. I think this paragraph would be even better if it had a sentence explaining where a user clicks in order to build their first Central Portal App.

Also, I lowercased CIAM because when things are turned into acronyms, that does not retroactively get the noun capitalized. The noun form is only capitalized if it is a proper noun. Customer identity and access management is a general concept, so not a proper noun.