Update how-to-set-up-your-partner-profile.md
The parter finder & profile has been completely revamped. This describes how to set up a partner profile in the new way. 

Btw: this is part of the Community Section, but I would like to have a dedicated Partner section as a main section in the documentation. How can this be done?