Update how-to-set-up-your-partner-profile.md
I would like to improve the answers that are provided by Maia chat. When I ask partner related questions, they are not aware of the partner finder. This document is a source that is used by Maia Chat. By extending this document, I expect Maia will give better answers. The FAQ section is a bit redundant for people, but it is highly valuable for the AI.